holy_mother_of_all_commit_bombs

+++ code/modules/library/soapstone.dm
@@ -6,7 +6,7 @@
 	throw_speed = 3
 	throw_range = 5
 	w_class = WEIGHT_CLASS_TINY
-	var/tool_speed = 50
+	var/tool_speed = 5 SECONDS
 	var/remaining_uses = 3
 
 /obj/item/soapstone/Initialize(mapload)
@@ -35,9 +35,12 @@
 		return
 
 	if(existing_message)
+		if(src in user.do_afters)
+			to_chat(user, "<span class='warning'>You're already trying to erase [existing_message]!</span>")
+			return
 		user.visible_message("<span class='notice'>[user] starts erasing [existing_message].</span>", "<span class='notice'>You start erasing [existing_message].</span>", "<span class='italics'>You hear a chipping sound.</span>")
 		playsound(loc, 'sound/items/gavel.ogg', 50, 1, -1)
-		if(do_after(user, tool_speed, target = existing_message))
+		if(do_after(user, tool_speed, target = existing_message, add_item = src))
 			user.visible_message("<span class='notice'>[user] erases [existing_message].</span>", "<span class='notice'>You erase [existing_message][existing_message.creator_key == user.ckey ? ", refunding a use" : ""].</span>")
 			existing_message.persists = FALSE
 			qdel(existing_message)
@@ -57,9 +60,12 @@
 	if(CHAT_FILTER_CHECK(message))
 		to_chat(user, "<span class='warning'>That message contains prohibited word(s)! Try again.</span>")
 		return
+	if(src in user.do_afters)
+		to_chat(user, "<span class='warning'>You're already engraving a message into [T]!</span>")
+
 	playsound(loc, 'sound/items/gavel.ogg', 50, 1, -1)
 	user.visible_message("<span class='notice'>[user] starts engraving a message into [T]...</span>", "<span class='notice'>You start engraving a message into [T]...</span>", "<span class='italics'>You hear a chipping sound.</span>")
-	if(can_use() && do_after(user, tool_speed, target = T) && can_use()) //This looks messy but it's actually really clever!
+	if(can_use() && do_after(user, tool_speed, target = T, add_item = src) && can_use()) //This looks messy but it's actually really clever!
 		if(!locate(/obj/structure/chisel_message) in T)
 			user.visible_message("<span class='notice'>[user] leaves a message for future spacemen!</span>", "<span class='notice'>You engrave a message into [T]!</span>", "<span class='italics'>You hear a chipping sound.</span>")
 			playsound(loc, 'sound/items/gavel.ogg', 50, 1, -1)
